requirements.txt
ccxt==4.3.41
huobi-client==0.3.1
huobi~=0.1.12
DateTime==5.5
aiohttp==3.9.5
requests==2.32.3
websocket-client==1.8.0
APScheduler~=3.10.4
redis==5.0.5
python-dotenv~=1.0.1
mysql-connector-python==8.4.0
huobipy~=1.0.1
pytz~=2024.1
python-binance==1.0.19
binance-futures-connector==4.0.0
binance==0.3
certifi==2024.6.2
cryptography==42.0.8
pyOpenSSL==24.1.0
urllib3~=2.2.1
psutil~=5.9.8